+++ source/blender/blenkernel/BKE_idcode.h
@@ -24,6 +24,8 @@
  * \ingroup bke
  */
 
+#include "BLI_sys_types.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -35,8 +37,8 @@ short BKE_idcode_from_name(const char *name);
 bool BKE_idcode_is_linkable(short idcode);
 bool BKE_idcode_is_valid(short idcode);
 
-int BKE_idcode_to_idfilter(const short idcode);
-short BKE_idcode_from_idfilter(const int idfilter);
+uint64_t BKE_idcode_to_idfilter(const short idcode);
+short BKE_idcode_from_idfilter(const uint64_t idfilter);
 
 int BKE_idcode_to_index(const short idcode);
 short BKE_idcode_from_index(const int index);
